+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Update the 'org.realityforge.arez' dependencies to version '0.103'
+
 ### [v0.27](https://github.com/arez/arez-promise/tree/v0.27) (2018-07-26)
 [Full Changelog](https://github.com/arez/arez-promise/compare/v0.26...v0.27)